pyproject.toml
[tool.poetry]
name = "victron-hass-mqtt"
version = "0.1.0"
description = "Python application for transmitting VE.Direct data to an mqtt server via the Home Assistant MQTT specification to enable automatic device discovery"
authors = ["Mathias Hedberg <mathias@hedberg.io>"]
license = "GPL-3.0-only"

[tool.poetry.scripts]
victron-hass-mqtt = "victron_hass_mqtt.__main__:main"

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"
PyYAML = "^6.0"
vedirect = {git = "https://github.com/karioja/vedirect.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
